Richtextを追加
