+++ src/heading/deprecated/schema.js
